requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-groq
langchain-ollama
faiss-cpu
chromadb
langchain-chroma
pypdf
ddgs
duckduckgo-search
numpy
scikit-learn
python-dotenv
transformers
groq
tiktoken
streamlit
requests
langgraph-checkpoint-sqlite